Fix build error: change MaxMsgSize type from long to int to match StreamConfig.MaxMsgSize

+++ SharpMUSH.Messaging/NATS/NatsOptions.cs
@@ -46,7 +46,7 @@ public class NatsOptions
 	/// Maximum size in bytes for a single message accepted by the JetStream stream.
 	/// Defaults to 6 MB. Set to -1 for unlimited.
 	/// </summary>
-	public long MaxMsgSize { get; set; } = 6 * 1024 * 1024;
+	public int MaxMsgSize { get; set; } = 6 * 1024 * 1024;
 
 	internal string GetConsumeStreamName() => ConsumeStreamName ?? StreamName;
 	internal string GetConsumeSubjectPrefix() => ConsumeSubjectPrefix ?? SubjectPrefix;
